today on things that never happened: Julian's story (probably) i mean he's flexin and it's cringe

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,18 +5,18 @@
  */
 void fib(void)
 {
-	unsigned long long fib1 = 1;
-	unsigned long long fib2 = 2;
-	unsigned long long fib3 = fib1 + fib2;
+	long fib1 = 1;
+	long fib2 = 2;
+	long fib3 = fib1 + fib2;
 	int i = 0;
 
-	printf("%llu, %llu, %llu", fib1, fib2, fib3);
+	printf("%lu, %lu, %lu", fib1, fib2, fib3);
 	for (i; i <= 50; i++)
 	{
 		fib1 = fib2;
 		fib2 = fib3;
 		fib3 = fib1 + fib2;
-		printf(", %llu", fib3);
+		printf(", %lu", fib3);
 	}
 	printf("\n");
 }
